requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
aniso8601==1.2.0          # via flask-restful
blinker==1.4              # via flask-mail, flask-principal
click==6.6                # via flask, flask-appconfig
flask-appconfig==0.11.1
flask-cors==3.0.2
Flask-Login==0.3.2        # via flask-security
Flask-Mail==0.9.1         # via flask-security
Flask-Principal==0.4.0    # via flask-security
flask-restful==0.3.5
flask-security==1.7.5
flask-sqlalchemy==2.1
Flask-WTF==0.13.1         # via flask-security
flask==0.11.1
itsdangerous==0.24        # via flask, flask-security
Jinja2==2.8               # via flask
jsl==0.2.4
MarkupSafe==0.23          # via jinja2
mod-wsgi==4.5.13          # Truncated Header Response workaround
passlib==1.6.5            # via flask-security
psycopg2==2.6.2
python-dateutil==2.5.3    # via aniso8601
pytz==2016.7              # via flask-restful
six==1.10.0               # via flask-appconfig, flask-cors, flask-restful, python-dateutil
SQLAlchemy==1.1.3         # via flask-sqlalchemy
Werkzeug==0.11.11         # via flask, flask-wtf
WTForms==2.1              # via flask-wtf
requests-futures==0.9.7   # added manually
